% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Format data for model fitting}
\usage{
build(survey1, survey2, survey3, paa.catch.female, paa.catch.male,
  n.trips.paa.catch, paa.survey1.female, paa.survey1.male, n.trips.paa.survey1,
  catch, paa.mature, weight.female, weight.male, misc.fixed.param = NULL,
  theta.ini = NULL, lkhd.paa = "normal", var.paa.add = TRUE,
  enable.priors = TRUE)
}
\arguments{
\item{survey1}{matrix (or data frame) of biomass estimates from survey 1 (g =
2); each row corresponds to a year of available data, the ﬁrst column
reports the year (T_2), the second column is the survey biomass estimate I
t,2 , and the third column is the observation error SD values κ t,2, for t
∈ T_2.}

\item{survey2}{matrix (or data frame) of biomass estimates from survey 2 (g =
3); each row corresponds to a year of available data, the ﬁrst column
reports the year (T_3), the second column is the survey biomass estimate I
t,3 , and the third column is the observation error SD values κ t,3, for t
∈ T_3.}

\item{survey3}{matrix (or data frame) of biomass estimates from survey 1 (g =
4); each row corresponds to a year of available data, the ﬁrst column
reports the year (T_4), the second column is the survey biomass estimate I
t,4 , and the third column is the observation error SD values κ t,4, for t
∈ T_4.}

\item{paa.catch.female}{matrix (or data frame) of paa for the females (s = 1)
from the commercial catch (g = 1); each row corresponds to a year of
available data, the ﬁrst column reports the year (U_1) and the next 30
columns correspond to the paa per age class and year p a,t,1,1 , for t ∈ U_1.}

\item{paa.catch.male}{matrix (or data frame) of paa for the males (s = 2)
from the commercial catch (g = 1); each row corresponds to a year of
available data, the ﬁrst column is the year (U_1) and the next 30 columns
correspond to the paa per age class and year p a,t,1,2 , for t ∈ U 1.}

\item{n.trips.paa.catch}{matrix (or data frame) of the number of trips per
year of available paa for the commercial catch (g = 1); each row
corresponds to a year, the ﬁrst column reports the year (U_1), the second
column is the number of trips n t,1 , for t ∈ U_1.}

\item{paa.survey1.female}{matrix (or data frame) of paa for the females
(s = 1) from survey 1 (g = 2); each row corresponds to a year of available data,
the ﬁrst column reports the year (U_2) and the next 30 columns correspond
to the paa per age class and year p a,t,2,1 , for t ∈ U_2.}

\item{paa.survey1.male}{matrix (or data frame) of paa for the males (s = 2)
from survey 1 (g = 2); each row corresponds to a year of available data,
the ﬁrst column is the year (U_2) and the next 30 columns correspond to
the paa per age class and year p a,t,2,2 , for t ∈ U_2.}

\item{n.trips.paa.survey1}{matrix (or data frame) of the number of trips per
year of available paa for survey 1 (g = 2); each row corresponds to a year,
the ﬁrst column reports the year (U_2), the second column is the number of
trips n t,2 , for t ∈ U 2.}

\item{catch}{Data frame (or matrix) of commercial trawl data, dimensions are
TC rows by 2 columns. The first column is the year, while the second column
is the observed catch.}

\item{paa.mature}{vector of A = 30 proportions of mature females.}

\item{weight.female}{vectors of A = 30 average weight at age of females (s = 1).}

\item{weight.male}{vectors of A = 30 average weight at age of males (s = 2).}

\item{misc.fixed.param}{optional named vector/list (or data frame) of ﬁxed
values for the selectivity parameters of surveys 2 and 3 and for the
process error SD σ R ; the names must match \code{muS2}, \code{deltaS2}, \code{upsilonS2}, \code{muS3},
\code{deltaS3}, \code{upsilonS3} and \code{sigmaR}. If unused (left to \code{NULL}), then the following
default values are used: muS2 = muS3 = 13.3, deltaS2 = deltaS3 = 0.22,
upsilonS2 = upsilonS3 = 10 and sigmaR = 0.05.}

\item{theta.ini}{optional named vector/list (or data frame) of starting
values for the model parameters to be estimated; the names must match \code{R0},
\code{M1}, \code{M2}, \code{muC}, \code{deltaC}, \code{upsilonC}, \code{muS1}, \code{deltaS1}, \code{upsilonS1}, \code{h}, \code{qS1}, \code{qS2} and
\code{qS3}. If unused (left to \code{NULL}), then the default values given in Table 2 are
used.}

\item{lkhd.paa}{either \code{"normal"} or \code{"binomial"}; speciﬁes a binomial likelihood
for the paa, following Cadigan et al. (2014), which may be more
statistically sound (and simpler) or a Gaussian likelihood as originally
speciﬁed in Edwards et al. (2014). Defaults to "normal".}

\item{var.paa.add}{either \code{TRUE} or \code{FALSE}; if \code{TRUE} (the default), it
enables the addition of the 1/(10A) term in Var\link{ξ a,t,g,s } related to the
Gaussian observation error of (F.19 ? ). If \code{FALSE}, this additional term
is 0 (disabled). If some observed p a,t,g,s are exactly 0 or 1, as in the
POP data, then it must be set to \code{TRUE} to avoid zero variances with
\code{lkhd.paa = "normal"}.}

\item{enable.priors}{either \code{TRUE} or \code{FALSE}; if \code{TRUE} (the default), it
enables all the prior distributions deﬁned in Section 3.1. If \code{FALSE}, the
priors are disabled and the maximization over θ is carried without any
constraints (apart from constraints on their range of possible values
enforced by means of transformations, such as strictly positive variances).}
}
\value{
A list of class \code{obj} properly formatted to be fed to \code{fit()},
with the following elements:
\itemize{
\item \code{datalist}: a list containing all DATA inputs for the TMB POP template;
\item \code{parlist}: a list containing all PARAMETER inputs for the TMB POP template;
}
}
\description{
Format data for model fitting
}
