% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{fit}
\alias{fit}
\title{Fits POP model to data supplied as POPobj}
\usage{
fit(obj, trace = TRUE, optim.control = list(eval.max = 5000, iter.max =
  5000))
}
\arguments{
\item{obj}{A list of class \code{obj} coming from either \code{build()}
(user-supplied data) or \code{sim()} (simulated data).}

\item{trace}{TODO}

\item{optim.control}{TODO}
}
\value{
A list with the following objects:
\itemize{
\item theta: vector of estimated fixed parameters
\item se.theta: vector of standard errors of fixed parameters
}
}
\description{
Fits POP model to data supplied as POPobj
}
