% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POPfit.r
\name{POPfit}
\alias{POPfit}
\title{Fits POP model to data supplied as POPobj}
\usage{
POPfit(POPobj, trace = TRUE, optim.control = list(eval.max = 5000,
  iter.max = 5000))
}
\arguments{
\item{POPobj}{x}

\item{trace}{x}

\item{optim.control}{x}

\item{POPobj}{A list of class POPobj coming from either POPbuild
(user-supplied data) or POPsim (simulated data).}
}
\value{
A list with the following objects:
\itemize{
  \item theta: vector of estimated fixed parameters;
  \item se.theta: vector of standard errors of fixed parameters;
}
}
\description{
to do later
}
\details{
Title

To be used with the POP cpp template.
}
