% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POPsim.r
\name{POPsim}
\alias{POPsim}
\title{Simulate data according to POP model.}
\usage{
POPsim(survey1, survey2, survey3, paa.catch.female, paa.catch.male,
  n.trips.paa.catch, paa.survey1.female, paa.survey1.male,
  n.trips.paa.survey1, catch, paa.mature, weight.female, weight.male,
  misc.fixed.param = NULL, theta.ini = NULL, lkhd.paa = "normal",
  var.paa.add = TRUE, enable.priors = TRUE)
}
\arguments{
\item{survey1}{x}

\item{survey2}{x}

\item{survey3}{x}

\item{paa.catch.female}{x}

\item{paa.catch.male}{x}

\item{n.trips.paa.catch}{x}

\item{paa.survey1.female}{x}

\item{paa.survey1.male}{x}

\item{n.trips.paa.survey1}{x}

\item{catch}{x}

\item{paa.mature}{x}

\item{weight.female}{x}

\item{weight.male}{x}

\item{misc.fixed.param}{x}

\item{theta.ini}{x}

\item{lkhd.paa}{x}

\item{var.paa.add}{x}

\item{enable.priors}{x}

\item{Ct}{Vector of length TC; total catch, fixed covariate.}
}
\value{
A list of class POPobj properly formatted to be fed to POPfit,
with the following objects:
\itemize{
  \item datalist: a list containing all DATA inputs for the TMB POP
  template;
  \item parlist: a list containing all PARAMETER inputs for the TMB POP
  template;
}
}
\description{
Uses fixed covariates and parameters to generate random effects and
response variables. Outputs a list of class "POPobj" that can be directly
supplied to \code{\link{POPfit}}.
}
\details{
Title

To be used with the POP cpp template. Compatible with v0.4.
}
