% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.r
\name{sim}
\alias{sim}
\title{Simulate data according to POP model.}
\usage{
sim(survey1, survey2, survey3, paa.catch.female, paa.catch.male,
  n.trips.paa.catch, paa.survey1.female, paa.survey1.male, n.trips.paa.survey1,
  catch, paa.mature, weight.female, weight.male, misc.fixed.param = NULL,
  theta.ini = NULL, lkhd.paa = "normal", var.paa.add = TRUE,
  enable.priors = TRUE)
}
\arguments{
\item{survey1}{TODO}

\item{survey2}{TODO}

\item{survey3}{TODO}

\item{paa.catch.female}{TODO}

\item{paa.catch.male}{TODO}

\item{n.trips.paa.catch}{TODO}

\item{paa.survey1.female}{TODO}

\item{paa.survey1.male}{TODO}

\item{n.trips.paa.survey1}{TODO}

\item{catch}{TODO}

\item{paa.mature}{TODO}

\item{weight.female}{TODO}

\item{weight.male}{TODO}

\item{misc.fixed.param}{TODO}

\item{theta.ini}{TODO}

\item{lkhd.paa}{TODO}

\item{var.paa.add}{TODO}

\item{enable.priors}{TODO}
}
\value{
A list of class POPobj properly formatted to be fed to \code{fit()},
with the following objects:
\itemize{
\item datalist: a list containing all DATA inputs for the TMB POP
template;
\item parlist: a list containing all PARAMETER inputs for the TMB POP
template;
}
}
\description{
Uses fixed covariates and parameters to generate random effects and
response variables. Outputs a list of class \code{obj} that can be directly
supplied to \code{fit()}.
}
\details{
To be used with the POP cpp template. Compatible with v0.4.
}
