% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POPbuild.r
\name{POPbuild}
\alias{POPbuild}
\title{Formats data files for POPfit.}
\usage{
POPbuild(survey1, survey2, survey3, paa.catch.female, paa.catch.male,
  n.trips.paa.catch, paa.survey1.female, paa.survey1.male,
  n.trips.paa.survey1, catch, paa.mature, weight.female, weight.male,
  misc.fixed.param = NULL, theta.ini = NULL, lkhd.paa = "normal",
  var.paa.add = TRUE, enable.priors = TRUE)
}
\arguments{
\item{survey1}{x}

\item{survey2}{x}

\item{survey3}{x}

\item{paa.catch.female}{x}

\item{paa.catch.male}{x}

\item{n.trips.paa.catch}{x}

\item{paa.survey1.female}{x}

\item{paa.survey1.male}{x}

\item{n.trips.paa.survey1}{x}

\item{catch}{x}

\item{paa.mature}{x}

\item{weight.female}{x}

\item{weight.male}{x}

\item{misc.fixed.param}{x}

\item{theta.ini}{x}

\item{lkhd.paa}{x}

\item{var.paa.add}{x}

\item{enable.priors}{x}

\item{catch}{Data frame (or matrix) of commercial trawl data, dimensions
are TC rows by 2 columns. The first column is the year, while the second
column is the observed catch.}
}
\value{
A list of class POPobj properly formatted to be fed to POPfit,
with the following objects:
\itemize{
  \item datalist: a list containing all DATA inputs for the TMB POP
  template;
  \item parlist: a list containing all PARAMETER inputs for the TMB POP
  template;
}
}
\description{
to do later
}
\details{
Title

To be used with the POP cpp template.
}
